+++ srcs/builtin/alias/manage_aliases.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   builtin_alias.c                                    :+:      :+:    :+:   */
+/*   manage_aliases.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jjaniec <jjaniec@student.42.fr>            +#+  +:+       +#+        */
+/*   By: cgaspart <cgaspart@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2018/10/07 15:06:12 by cyfermie          #+#    #+#             */
-/*   Updated: 2018/11/13 18:58:38 by jjaniec          ###   ########.fr       */
+/*   Created: 2018/12/10 17:10:30 by cgaspart          #+#    #+#             */
+/*   Updated: 2018/12/10 17:19:49 by cgaspart         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,10 +23,7 @@ static bool			save_aliases_in_file(struct s_alias *alias)
 	&& write(2, SH_NAME": error with file ."SH_NAME"_aliases\n", 36))
 		return (false);
 	if (alias->key == NULL)
-	{
-		close(fd);
-		return (true);
-	}
+		return (close(fd) ? (true) : (true));
 	while (alias != NULL)
 	{
 		if (ft_strlen(alias->key) > 0U)
@@ -98,31 +95,6 @@ static bool			add_tmp_alias(const char *key, const char *value, \
 	return (true);
 }
 
-static void			print_aliases(struct s_alias *alias)
-{
-	if (alias->key == NULL && "list is empty")
-		return ;
-	while (alias != NULL)
-	{
-		if (ft_strlen(alias->key) > 0U)
-			ft_printf("'%s' = '%s'\n", alias->key, alias->value);
-		alias = alias->next;
-	}
-}
-
-static bool			check_option(const char *arg)
-{
-	if (ft_strequ(arg, "-d"))
-		return (true);
-	if (arg[0] == '-')
-		return (false);
-	while (*arg != '\0' && is_separator(*arg))
-		++arg;
-	if (*arg == '-')
-		return (false);
-	return (true);
-}
-
 static void			delete_alias(const char *key, struct s_alias *alias)
 {
 	if (alias->key == NULL && "list is empty")
@@ -141,38 +113,13 @@ static void			delete_alias(const char *key, struct s_alias *alias)
 	}
 }
 
-/*
-**	$ alias key value  => create a new entry
-**	$ alias --save  => save the aliases list into a file
-**	$ alias -d key  => delete an entry
-**	$ alias  => display the aliases list
-*/
-
-void				builtin_alias(char **argv, t_environ *env, t_exec *exe)
+void				manage_aliases(char **argv, struct s_alias *alias,
+									t_exec *exe, unsigned int nb_args)
 {
-	unsigned int	nb_args;
-	struct s_alias	*alias;
-
-	(void)env;
-	exe->ret = 0;
-	alias = access_alias_datas();
-	nb_args = ft_count_elem_2d_array(argv + 1);
-	if ((nb_args != 0 && nb_args != 1 && nb_args != 2) \
-	|| (nb_args == 2 && check_option(argv[1]) == false))
+	if (nb_args == 1)
 	{
-		ft_putstr_fd(BUILTIN_ALIAS_USAGE, STDERR_FILENO);
-		exe->ret = 1;
-		return ;
-	}
-	if (nb_args == 0)
-		print_aliases(alias);
-	else if (nb_args == 1)
-	{
-		if (!ft_strequ("--save", argv[1]))
-		{
+		if (!ft_strequ("--save", argv[1]) && (exe->ret = 1))
 			ft_putstr_fd(BUILTIN_ALIAS_USAGE, STDERR_FILENO);
-			exe->ret = 1;
-		}
 		else
 			exe->ret = save_aliases_in_file(alias) == true ? (0) : (1);
 	}
