+++ srcs/builtin/builtin_setenv.c
@@ -6,14 +6,15 @@
 /*   By: jjaniec <jjaniec@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/09/27 19:34:40 by jjaniec           #+#    #+#             */
-/*   Updated: 2018/11/11 19:58:18 by jjaniec          ###   ########.fr       */
+/*   Updated: 2018/12/10 18:03:49 by cgaspart         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <forty_two_sh.h>
 
 /*
-** Check args format and return 1 to print error messages if any of them is invalid
+** Check args format and return 1 to print error messages
+** if any of them is invalid
 */
 
 static int	check_args(char **argv)
@@ -25,7 +26,8 @@ static int	check_args(char **argv)
 			return (1);
 		else if (is_identifier_invalid(*argv, assign_ptr))
 			return (2);
-		else argv++;
+		else
+			argv++;
 	return (0);
 }
 
@@ -41,6 +43,19 @@ static void	print_setenv_error(int err)
 		ft_putstr_fd(SH_NAME": invalid identifiers\n", 2);
 }
 
+static void	add_setenv(t_environ *origin_env, char **ptr)
+{
+	while (ptr && *ptr)
+	{
+		if (!(origin_env->get_var(origin_env, *ptr)))
+			origin_env->add_var(origin_env, *ptr, NULL);
+		else
+			ft_strncpy(origin_env->last_used_elem->entry, *ptr,
+			MAX_ENV_ENTRY_LEN);
+		ptr++;
+	}
+}
+
 /*
 ** Add environnement variables in our t_env_entry linked list w/ add_var
 */
@@ -51,9 +66,6 @@ void		builtin_setenv(char **argv, t_environ *env, t_exec *exe)
 	int			err;
 	t_environ	*origin_env;
 
-	(void)argv;
-	(void)env;
-	(void)exe;
 	ptr = argv + 1;
 	if (!(argv[1]))
 	{
@@ -67,13 +79,6 @@ void		builtin_setenv(char **argv, t_environ *env, t_exec *exe)
 		exe->ret = 1;
 		return ;
 	}
-	while (ptr && *ptr)
-	{
-		if (!(origin_env->get_var(origin_env, *ptr)))
-			origin_env->add_var(origin_env, *ptr, NULL);
-		else
-			ft_strncpy(origin_env->last_used_elem->entry, *ptr, MAX_ENV_ENTRY_LEN);
-		ptr++;
-	}
+	add_setenv(origin_env, ptr);
 	exe->ret = 0;
 }
